+++ Products/GenericSetup/tool.py
@@ -1264,6 +1264,14 @@ class SetupTool(Folder):
         # got passed the profile_id.
         last_num = len(chain)
         for num, profile_id in enumerate(chain, 1):
+            # Purge profile versions when this is a base profile.
+            try:
+                profile_type = self.getProfileInfo(profile_id).get('type')
+            except KeyError:
+                # this will be a snapshot profile
+                profile_type = None
+            if profile_type == BASE:
+                self.purgeProfileVersions()
             if num == last_num:
                 generic_logger.info('Applying main profile %s', profile_id)
             else: